requirements/requirements.txt
os-scrapy>=0.0.4
